+++ internal/cmd/context/use.go
@@ -39,6 +39,17 @@ func UseCmd(opts *clioptions.CLIOptions) *cobra.Command {
 			fmt.Printf("Switched to context \"%s\"\n", newContext)
 			return nil
 		},
+		ValidArgsFunction: func(_ *cobra.Command, args []string, _ string) ([]string, cobra.ShellCompDirective) {
+			if len(args) > 0 {
+				return nil, cobra.ShellCompDirectiveNoFileComp
+			}
+			locator := cliconfig.NewConfigPathLocator()
+			config, err := locator.ReadConfig()
+			if err != nil {
+				return nil, cobra.ShellCompDirectiveNoFileComp
+			}
+			return listContexts(config), cobra.ShellCompDirectiveNoFileComp
+		},
 	}
 
 	return cmd