+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `extensions delete` command
 - `extensions activate` command
 - `extensions deactivate` command
+- `context use` autocomplete contexts with tab
 
 ### Changed