+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.13.0] - 2024-06-26
+
 ### Added
 
 - `extensions list` command
@@ -263,7 +265,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - create cli sdk
 - create cli renderer
 
-[unreleased]: https://github.com/mia-platform/miactl/compare/v0.12.2...HEAD
+[unreleased]: https://github.com/mia-platform/miactl/compare/v0.13.0...HEAD
+[0.13.0]: https://github.com/mia-platform/miactl/compare/v0.12.2...v0.13.0
 [0.12.2]: https://github.com/mia-platform/miactl/compare/v0.12.1...v0.12.2
 [0.12.1]: https://github.com/mia-platform/miactl/compare/v0.12.0...v0.12.1
 [0.12.0]: https://github.com/mia-platform/miactl/compare/v0.11.0...v0.12.0